streamlit/requirements.txt
streamlit>=1.30.0
snowflake-snowpark-python>=1.11.0
snowflake-ml-python>=1.2.0
pandas>=2.0.0
plotly>=5.18.0
